convert pow_tests.cpp
