+++ src/Makefile.test.include
@@ -46,10 +46,6 @@ BITCOIN_TESTS =\
   test/checkblock_tests.cpp \
   test/coins_tests.cpp \
   test/dbwrapper_tests.cpp \
-  test/main_tests.cpp \
-  test/mruset_tests.cpp \
-  test/pmt_tests.cpp \
-  test/pow_tests.cpp \
   test/prevector_tests.cpp \
   test/reverselock_tests.cpp \
   test/rpc_tests.cpp \
